Moved LocalLambdaRunner to its own file
